Update EpochToDateTimeConverter.cs
I need to serialize the webhook objects, but I am unable to. I cannot use XmlSerializer because there are ILists in the code which is an interface that the XmlSerializer cannot handle, and I cannot use JsonConvert because you have overridden the DateTime converter which throws a NotImplementedException when I try to serialize. 
By just calling the basic functionality for serializing DateTime instead of throwing an exception, it should work fine.